requirements.txt
pyserial==3.5.0
Pillow==10.0.0
heatshrink2==0.12.0